multi-storage-client/pyproject.toml
[build-system]
requires = [
    "maturin>=1.9.1,<2"
]
build-backend = "maturin"

[project]
name = "multi-storage-client"
version = "0.33.0"
description = "Unified high-performance Python client for object and file stores."
authors = [
    { name = "NVIDIA Multi-Storage Client Team" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.9,<4"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]
dependencies = [
    "filelock>=3.14,<4",
    "jmespath>=1.0.1,<2",
    "jsonschema>=4,<5",
    "python-dateutil>=2.8,<3",
    "pyyaml>=6.0.1,<7",
    "opentelemetry-api>=1.24,<2",
    "xattr>=1.1.4,<2",
    "wcmatch>=10,<11",
    "tqdm>=4.67.1,<5",
    "lark>=1.2.2,<2",
    "prettytable>=3.10,<4",
    "psutil>=7,<8",
    "tzdata>=2024.2,<2026",
]

[project.urls]
source = "https://github.com/NVIDIA/multi-storage-client"
documentation = "https://nvidia.github.io/multi-storage-client/"

[project.optional-dependencies]
# Observability.
observability-otel = [
    "opentelemetry-sdk>=1.24,<2",
    "opentelemetry-exporter-otlp-proto-http>=1.24,<2",
]
msal = [
    "msal>=1.31.1,<2"
]

# MCP.
mcp = [
    "fastmcp>=2.11.2; python_version>='3.10'"
]


# Storage providers.
aistore = [
    "aistore>=1.16,<2"
]
azure-storage-blob = [
    "azure-storage-blob>=12.17,<13"
]
boto3 = [
    "boto3>=1.36,<2"
]
google-cloud-storage = [
    "google-cloud-storage>=3.2,<4"
]
huggingface = [
    "huggingface_hub>=0.34.4,<1",
    "hf_transfer>=0.1.9"
]
oci = [
    "oci>=2.150,<3"
]
parquet = [
    "pyarrow>=21.0"
]

# Higher-level libraries.
fsspec = [
    "fsspec>=2023.9.0"
]
hydra-core = [
    "hydra-core>=1.3,<2"
]
numpy = [
    "numpy>=2,<2.1; python_version == '3.9'",
    "numpy>=2.1,<3; python_version >= '3.10'"
]
ray = [
    "ray>=2.48,<3"
]
torch = [
    "torch>=2.8,<3"
]
xarray = [
    "xarray>=2024.07.0"
]
zarr = [
    "zarr>=2.18.2,<3; python_version == '3.9'",
    "zarr>=2.18.3,<3; python_version == '3.10'",
    "zarr>=2.18.7,<3; python_version >= '3.11'"
]

[dependency-groups]
dev = [
    "types-PyYAML>=6.0.1,<7",
    "types-python-dateutil>=2.8,<3",
    "types-requests>=2.25,<3",
    "maturin[zig]>=1.9.1,<2",
    "pytest>=8.4.1,<9",
    "pytest-asyncio>=1.1,<2",
    "pytest-cov>=6.2.1,<7",
    "pytest-forked>=1.6,<2",
    "pytest-stress>=1.0.1,<2",
    "pytest-timeout>=2.4,<3",
    "pytest-xdist[psutil]>=3.8,<4",
    "coverage[toml]>=7.9.2,<8"
]

[project.scripts]
msc-benchmark = "multistorageclient.commands.msc_benchmark:main"
msc = "multistorageclient.commands.cli.main:main"

[project.entry-points."fsspec.specs"]
msc = "multistorageclient.contrib.async_fs.MultiStorageAsyncFileSystem"

[tool.coverage.run]
branch = true
concurrency = [
    "multiprocessing",
    "thread"
]
data_file = ".reports/unit/.coverage"
parallel = true
source = [
    "src"
]

[tool.coverage.html]
directory = ".reports/unit/coverage"

[tool.coverage.xml]
output = ".reports/unit/coverage.xml"

[tool.maturin]
features = [
    "pyo3/extension-module"
]
manifest-path = "rust/Cargo.toml"
python-packages = [
    "multistorageclient",
    "multistorageclient_rust",
    "hydra_plugins"
]

[tool.pyright]
include = [
    "src",
    "tests"
]
typeCheckingMode = "basic"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib"
]
asyncio_default_fixture_loop_scope = "function"
markers = [
    "serial: marks tests as serial (deselect with '-m \"not serial\"')"
]
pythonpath = [
    "tests"
]
testpaths = [
    "tests/test_multistorageclient/unit",
    "tests/test_multistorageclient_rust/unit"
]
timeout = 600

[tool.ruff]
extend = "../pyproject.toml"
